chore(drawer): rename drawer-animation file
