+++ projects/extensions/drawer/public-api.ts
@@ -2,5 +2,5 @@ export * from './drawer-module';
 export * from './drawer';
 export * from './drawer-config';
 export * from './drawer-container';
-export * from './drawer-animation';
+export * from './drawer-animations';
 export * from './drawer-ref';
